refactor setup

+++ python/setup_python.sh
@@ -7,15 +7,21 @@ export PATH="/usr/local/opt/curl/bin:$PATH"
 
 pyenv_install_python(){
 	version=$1
-	if [ ! -e ${PYENV_ROOT}/versions/${version}/bin/pip ]; then
-        echo "Error: pip is not installed for python versions: ${version}"
-        echo "Setting up pip for version ${version}"
-        pyenv local $version
-		wget https://bootstrap.pypa.io/get-pip.py
-		python get-pip.py
-		pip install dotfiles
-		rm get-pip.py
-    fi 	
+
+	fact=`pyenv versions | grep $version`
+	if [ "x${fact}" == "x" ]; then 
+		echo "Pyenv : not installed ... ${version}"
+	else
+		if [ ! -e ${PYENV_ROOT}/versions/${version}/bin/pip ]; then
+	        echo "Error: pip is not installed for python versions: ${version}"
+	        echo "Setting up pip for version ${version}"
+	        pyenv local $version
+			wget https://bootstrap.pypa.io/get-pip.py
+			python get-pip.py
+			pip install dotfiles
+			rm get-pip.py
+	    fi 
+    fi	
 }
 
 # Keep 2.7.14 as default python version
